LayoutTests/imported/w3c/web-platform-tests/dom/traversal/TreeWalker-acceptNode-filter-cross-realm-expected.txt



PASS NodeFilter is cross-realm plain object without 'acceptNode' property
PASS NodeFilter is cross-realm plain object with non-callable 'acceptNode' property
FAIL NodeFilter is cross-realm plain object with revoked Proxy as 'acceptNode' property assert_throws_js: function "() => { walker.firstChild(); }" threw object "TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it" ("TypeError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS NodeFilter is cross-realm non-callable revoked Proxy
PASS NodeFilter is cross-realm callable revoked Proxy
